% Auto-generated: do not edit by hand
\name{''StorageObserver}

\alias{''StorageObserver}

\title{StorageObserver component}

\description{

}

\usage{
''StorageObserver(id=NULL, clear=NULL, interval=NULL, limit=NULL)
}

\arguments{
\item{id}{Character. }

\item{clear}{Logical. }

\item{interval}{Numeric. }

\item{limit}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

